convert calendarEventTooltip utility fn into service.
this allows us to inject the intl service directly into the tooltip
service, rather than passing it as input to the util function.
which will help with tracking and linting translations used within.
