+++ packages/ilios-common/app/services/calendar-event-tooltip.js
@@ -1,0 +1,1 @@
+export { default } from 'ilios-common/services/calendar-event-tooltip';
