--- packages/ilios-common/app/utils/calendar-event-tooltip.js
@@ -1,1 +1,0 @@
-export { default } from 'ilios-common/utils/calendar-event-tooltip';
